requirements.txt
Flask
groq
python-dotenv
werkzeug
Pillow
PyPDF2
pytesseract
easyocr